python-functions/requirements.txt
azure-functions
requests
beautifulsoup4
python-dateutil
email-validator